+++ Editor/AnimatorParserV2/AnimationParser.cs
@@ -49,10 +49,10 @@ namespace Anatawa12.AvatarOptimizer.AnimatorParsersV2
             var children = blendTree.children;
 
             return NodesMerger.Merge<
-                ImmutableNodeContainer, ImmutablePropModNode<float>, ImmutablePropModNode<Object>,
-                BlendTreeElement<float>, BlendTreeElement<Object>,
-                ImmutableNodeContainer, ImmutableNodeContainer, ImmutablePropModNode<float>, 
-                ImmutablePropModNode<Object>,
+                ImmutableNodeContainer, ImmutablePropModNode<FloatValueInfo>, ImmutablePropModNode<ObjectValueInfo>,
+                BlendTreeElement<FloatValueInfo>, BlendTreeElement<ObjectValueInfo>,
+                ImmutableNodeContainer, ImmutableNodeContainer, ImmutablePropModNode<FloatValueInfo>, 
+                ImmutablePropModNode<ObjectValueInfo>,
                 BlendTreeMergeProperty
             >(children.Select(x => ParseMotionInner(root, x.motion, mapping)),
                 new BlendTreeMergeProperty(blendTree.blendType));
@@ -60,10 +60,10 @@ namespace Anatawa12.AvatarOptimizer.AnimatorParsersV2
 
         internal readonly struct BlendTreeMergeProperty : 
             IMergeProperty1<
-                ImmutableNodeContainer, ImmutablePropModNode<float>, ImmutablePropModNode<Object>,
-                BlendTreeElement<float>, BlendTreeElement<Object>,
-                ImmutableNodeContainer, ImmutableNodeContainer, ImmutablePropModNode<float>,
-                ImmutablePropModNode<Object>
+                ImmutableNodeContainer, ImmutablePropModNode<FloatValueInfo>, ImmutablePropModNode<ObjectValueInfo>,
+                BlendTreeElement<FloatValueInfo>, BlendTreeElement<ObjectValueInfo>,
+                ImmutableNodeContainer, ImmutableNodeContainer, ImmutablePropModNode<FloatValueInfo>,
+                ImmutablePropModNode<ObjectValueInfo>
             >
         {
             private readonly BlendTreeType _blendType;
@@ -76,17 +76,17 @@ namespace Anatawa12.AvatarOptimizer.AnimatorParsersV2
             public ImmutableNodeContainer CreateContainer() => new();
             public ImmutableNodeContainer GetContainer(ImmutableNodeContainer source) => source;
 
-            public BlendTreeElement<float> GetIntermediate(ImmutableNodeContainer source,
-                ImmutablePropModNode<float> node, int index) => new BlendTreeElement<float>(index, node);
+            public BlendTreeElement<FloatValueInfo> GetIntermediate(ImmutableNodeContainer source,
+                ImmutablePropModNode<FloatValueInfo> node, int index) => new BlendTreeElement<FloatValueInfo>(index, node);
 
-            public BlendTreeElement<Object> GetIntermediate(ImmutableNodeContainer source,
-                ImmutablePropModNode<Object> node, int index) => new BlendTreeElement<Object>(index, node);
+            public BlendTreeElement<ObjectValueInfo> GetIntermediate(ImmutableNodeContainer source,
+                ImmutablePropModNode<ObjectValueInfo> node, int index) => new BlendTreeElement<ObjectValueInfo>(index, node);
 
-            public ImmutablePropModNode<float> MergeNode(List<BlendTreeElement<float>> nodes, int sourceCount) =>
-                new BlendTreeNode<float>(nodes, _blendType, partial: nodes.Count != sourceCount);
+            public ImmutablePropModNode<FloatValueInfo> MergeNode(List<BlendTreeElement<FloatValueInfo>> nodes, int sourceCount) =>
+                new BlendTreeNode<FloatValueInfo>(nodes, _blendType, partial: nodes.Count != sourceCount);
 
-            public ImmutablePropModNode<Object> MergeNode(List<BlendTreeElement<Object>> nodes, int sourceCount) =>
-                new BlendTreeNode<Object>(nodes, _blendType, partial: nodes.Count != sourceCount);
+            public ImmutablePropModNode<ObjectValueInfo> MergeNode(List<BlendTreeElement<ObjectValueInfo>> nodes, int sourceCount) =>
+                new BlendTreeNode<ObjectValueInfo>(nodes, _blendType, partial: nodes.Count != sourceCount);
         }
 
         private readonly Dictionary<(GameObject, AnimationClip), ImmutableNodeContainer> _parsedAnimationCache = new();